Move Google Reader API routes to standard paths
Apps like Read You and FeedMe expect the Google Reader API at
/accounts/ClientLogin for auth and /reader/api/0/* for API endpoints.
Our routes were at /api/reader/api/0/* which meant no single server
URL would work for clients.

Move routes to match the standard layout:
- /accounts/ClientLogin (auth)
- /reader/api/0/* (all API endpoints)

This lets users enter their base URL (e.g. https://example.com)
and have clients work correctly out of the box.
