+++ src/app/demo/articles/google-reader-api.ts
@@ -57,7 +57,7 @@ const article: DemoArticle = {
 
     <h3>How It Works</h3>
 
-    <p>The Google Reader API is implemented as a set of Next.js route handlers under <code>/api/reader/api/0/</code>. Rather than duplicating business logic, each endpoint is a thin translation layer that converts between the Google Reader wire format and Lion Reader&rsquo;s existing services layer. This means behavior is identical whether you&rsquo;re reading through the web UI, the <a href="/demo/all?entry=mcp-server">MCP server</a>, or a third-party app.</p>
+    <p>The Google Reader API is implemented as a set of Next.js route handlers under <code>/reader/api/0/</code> (with authentication at <code>/accounts/ClientLogin</code>). Rather than duplicating business logic, each endpoint is a thin translation layer that converts between the Google Reader wire format and Lion Reader&rsquo;s existing services layer. This means behavior is identical whether you&rsquo;re reading through the web UI, the <a href="/demo/all?entry=mcp-server">MCP server</a>, or a third-party app.</p>
 
     <p>One interesting challenge is ID mapping. Google Reader clients expect signed 64-bit integer IDs, but Lion Reader uses UUIDv7 (128-bit). The API derives a deterministic 63-bit integer from each UUID by extracting the 48-bit timestamp and 15 bits of randomness. This preserves time-ordering (so clients sort correctly) and is fully reversible without any extra storage.</p>
   `,
